refactor(backend): divide questoes
faz a divisão das responsabilidade das questoes
