+++ projeto/src/routes/questoes.routes.js
@@ -1,189 +1,24 @@
 const { Router } = require("express");
 const authMiddleware = require("../middlewares/auth.middleware");
 const {
-    findProximaQuestaoByUsuario,
-    findQuestaoDoExameByUsuario,
-    findRespostaByExameEQuestao,
-    inserirRespostaQuestao,
-    usuarioConcluiuModuloAtual,
-    findModuloAtualByUsuario,
-    findOutroGrupoAleatorio,
-    updateProximaTentativa,
-    findProximoModuloByUsuario,
-    updateProximoModulo,
-    findModulosRespondidosByUsuario
-} = require("../repositories/questoes.repository");
+    getProximaQuestaoController,
+    responderQuestaoController,
+    proximaTentativaController,
+    proximoModuloController,
+    getModulosRespondidosController,
+} = require("../controllers/questoes.controller");
 
 const router = Router();
 
-router.get("/proxima-questao", authMiddleware, async function (req, res) {
-  try {
-    const questao = await findProximaQuestaoByUsuario(req.usuario.id_usuario);
+router.get("/proxima-questao", authMiddleware, getProximaQuestaoController);
 
-    if (!questao) {
-      return res
-        .status(404)
-        .json({ message: "Nenhuma questão pendente encontrada" });
-    }
+router.post("/responder", authMiddleware, responderQuestaoController);
 
-    return res.status(200).json({
-    ...questao,
-    imagem: questao.imagem ? `/imagens/questoes/${questao.imagem}` : null,
-  });
-  } catch (e) {
-    return res.status(500).json({
-      message: "Erro interno do servidor",
-    });
-  }
-});
+router.patch("/proxima-tentativa", authMiddleware, proximaTentativaController);
 
-router.post("/responder", authMiddleware, async function (req, res) {
-  try {
-    console.log("body", req.body);
-    const { id_exame, id_questao, resposta } = req.body;
+router.patch("/proximo-modulo", authMiddleware, proximoModuloController);
 
-    const respostaNormalizada = resposta.trim().toLowerCase();
-
-    const questao = await findQuestaoDoExameByUsuario(req.usuario.id_usuario, id_exame, id_questao);
-
-    if (!questao) {
-      return res.status(404).json({
-        message: "Questão não encontrada para este exame",
-      });
-    }
-
-    const respostaExistente = await findRespostaByExameEQuestao(
-      id_exame,
-      id_questao,
-);
-
-    if (respostaExistente) {
-      return res.status(409).json({
-        message: "Questão já respondida",
-      });
-    }
-
-    const nota = questao.alternativa_correta === respostaNormalizada ? 1 : 0;
-    const respostaInserida = await inserirRespostaQuestao(id_exame, id_questao, respostaNormalizada,nota);
-
-    return res.status(201).json(respostaInserida);
-  } catch (e) {
-    return res.status(500).json({
-      message: "Erro interno do servidor",
-    });
-  }
-});
-
-router.patch("/proxima-tentativa", authMiddleware, async function (req, res) {
-  try {
-    const concluido = await usuarioConcluiuModuloAtual(req.usuario.id_usuario);
-    if (!concluido) {
-      return res.status(409).json({
-        message: "Você ainda não concluiu todas as questões do módulo atual",
-      });
-    }
-
-    const modulo = await findModuloAtualByUsuario(req.usuario.id_usuario);
-    if (!modulo) {
-      return res.status(404).json({
-        message: "Módulo atual não encontrado",
-      });
-    }
-    if (modulo.tentativa >= 2) {
-      return res.status(409).json({
-        message: "Limite de 2 tentativas atingido",
-      });
-    }
-
-    const grupo = await findOutroGrupoAleatorio(
-      req.usuario.id_usuario,
-      modulo.id_modulo,
-    );
-    if (!grupo) {
-      return res.status(404).json({
-        message: "Nenhum grupo alternativo disponível para este módulo",
-      });
-    }
-
-    const exame = await updateProximaTentativa(
-      modulo.id_exame,
-      grupo,
-      modulo.tentativa + 1,
-    );
-    if (!exame) {
-      return res.status(404).json({
-        message: "Exame não encontrado para atualização",
-      });
-    }
-
-    return res.status(200).json(exame);
-  } catch (e) {
-    console.log(e.message)
-    return res.status(500).json({
-      message: "Erro interno do servidor",
-    });
-  }
-});
-
-router.patch("/proximo-modulo", authMiddleware, async function (req, res) {
-  try {
-    const concluido = await usuarioConcluiuModuloAtual(req.usuario.id_usuario);
-    if (!concluido) {
-      return res.status(409).json({
-        message: "Você ainda não concluiu todas as questões do módulo atual",
-      });
-    }
-
-    const moduloAtual = await findModuloAtualByUsuario(req.usuario.id_usuario);
-    if (!moduloAtual) {
-      return res.status(404).json({
-        message: "Módulo atual não encontrado",
-      });
-    }
-
-
-    const modulo = await findProximoModuloByUsuario(req.usuario.id_usuario);
-    if (!modulo) {
-      return res.status(404).json({
-        message: "Você concluiu todos os módulos",
-      });
-    }
-
-
-    const grupo = await findOutroGrupoAleatorio(req.usuario.id_usuario, modulo);
-    if( !grupo ){
-      return res.status(404).json({
-        message: "Nenhum grupo disponível para o próximo módulo",
-      });
-    }
-    console.log("Grupo",grupo);
-
-    const exame = await updateProximoModulo(moduloAtual.id_exame, modulo, grupo, 1);
-    if (!exame) {
-      return res.status(404).json({
-        message: "Exame não encontrado para atualização",
-      });
-    }
-
-    return res.status(200).json(exame);
-  } catch (e) {
-    return res.status(500).json({
-      message: "Erro interno do servidor",
-    });
-  }
-});
-
-router.get("/modulos-respondidos", authMiddleware, async function (req, res) {
-  try {
-    const modulos = await findModulosRespondidosByUsuario(req.usuario.id_usuario);
-
-    return res.status(200).json(modulos);
-  } catch (e) {
-    return res.status(500).json({
-      message: "Erro interno do servidor",
-    });
-  }
-});
+router.get("/modulos-respondidos", authMiddleware, getModulosRespondidosController);
 
 /*
 ---- Comandos CURL para questões ----
